encounter tweaks
