+++ client/src/pages/Campaign.js
@@ -4,6 +4,7 @@ import { EncounterTracker } from '../components/EncounterTracker';
 import CharacterList from '../components/CharacterList';
 import { Section } from '../components/Section';
 import CharacterSheet from '../components/CharacterSheet';
+import EncounterForm from '../components/EncounterForm';
 import { Link, useParams } from 'react-router-dom';
 import { useQuery } from '@apollo/client';
 import { QUERY_CAMPAIGN } from '../utils/queries';
@@ -22,6 +23,9 @@ function Campaign() {
 
   const campaign = campaignData?.campaign || {};
   const [chars, setChars] = useState([]);
+  const [activeEncounter, setActiveEncounter] = useState({});
+  const [encounterFormOpen, setEncouterFormOpen] = useState(false);
+
   useEffect(() => {
     campaignData?.campaign && setChars(campaignData.campaign.characters);
   }, [campaignData]);
@@ -34,12 +38,16 @@ function Campaign() {
           <CharacterList chars={chars} setChars={setChars} />
         </Section>
         <Section title='Encounter List' collapsable>
-          {campaign?.encounters?.map(enc => (
-            <div>{enc.title}</div>
+          <button>New Encounter</button>
+          {campaign?.encounters?.map((enc, i) => (
+            <div key={i} onClick={() => setActiveEncounter(enc)}>
+              <div>{enc.title}</div>
+              {/* <div>{JSON.stringify(enc)}</div> */}
+            </div>
           ))}
         </Section>
-        <Section title='Encounter Tracker' collapsable>
-          <EncounterTracker chars={chars} />
+        <Section title='Encounter Tracker' collapsable startOpen={false}>
+          <EncounterTracker chars={chars} activeEncounter={activeEncounter} />
         </Section>
         <Section title='Character Sheet' collapsable startOpen={false}>
           {chars.length > 0 && (
@@ -47,6 +55,7 @@ function Campaign() {
           )}
         </Section>
       </section>
+      <EncounterForm setActiveEncounter={setActiveEncounter} chars={chars} />
     </PageWrapper>
   );
 }
